scripts/cangen/pyproject.toml
[tool.poetry]
name = "cangen"
version = "0.1.0"
description = "Converts .dbc files into C++ files for a CAN bus system."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
argparse-addons = "0.12.0"
bitstruct = "8.19.0"
cantools = "39.4.5"
crccheck = "1.3.0"
diskcache = "5.6.3"
Jinja2 = "3.1.3"
MarkupSafe = "2.1.5"
numpy = "1.26.4"
packaging = "24.0"
python-can = "4.3.1"
PyYAML = "6.0.1"
textparser = "0.24.0"
typing_extensions = "4.11.0"
wrapt = "1.16.0"

[tool.poetry.scripts]
cangen = "cangen.main:main"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
